解决掉一卡一卡的Bug

+++ QEABoatServer/src/main/StatusMachine.java
@@ -1,8 +1,8 @@
 package main;
 
 public class StatusMachine {
-	public static int upFlags = 0;
-	public static int downFlags = 0;
-	public static int leftFlags = 0;
-	public static int rightFlags = 0;
+	public static long upFlags = 0;
+	public static long downFlags = 0;
+	public static long leftFlags = 0;
+	public static long rightFlags = 0;
 }
